make it better

+++ src/components/diccionario/services/index.js
@@ -1,15 +1,27 @@
 import axios from 'axios'
+import definicionesJSON from "./definiciones.json"
 //const baseURL = process.env.REACT_APP_BASE_URL
 const baseURL = "http://www.diccionarioprogramacion.somee.com/api";
 
+const ModoProductivo = false;
+
 export async function getDefiniciones(){
         try{
-            const respuesta = await axios({
-                url: `${baseURL}/definiciones`,
-                method: 'GET'
-            })
 
+            if(ModoProductivo){
+                const respuesta = await axios({
+                    url: `${baseURL}/definiciones`,
+                    method: 'GET'
+                })
+            
             return respuesta;
+            
+            }else{
+                let defJSON = JSON.stringify(definicionesJSON);
+                let defObj = JSON.parse(defJSON);
+
+                return defObj;
+            }
 
         }catch(e){
             console.log(e)
@@ -18,12 +30,20 @@ export async function getDefiniciones(){
 
 export async function getDataDefinicion(id){
     try{
-        const respuesta = await axios({
-            url: `${baseURL}/definiciones/${id}`,
-            method: 'GET'
-        })
-        console.log("wat"+respuesta);
-        return respuesta;
+
+        if(ModoProductivo){
+            const respuesta = await axios({
+                url: `${baseURL}/definiciones/${id}`,
+                method: 'GET'
+            })
+
+            return respuesta;
+
+        }else{
+            let defFind = definicionesJSON.findIndex((def) => def.id === id);
+            return defFind;
+        }
+       
     }catch(e){
             console.log(e)
     }
@@ -32,17 +52,21 @@ export async function getDataDefinicion(id){
 export async function guardaDefinicion(definicionData){
     try{
 
-        const formData = new FormData();
-        formData.append('palabra', definicionData.palabra);
-        formData.append('definicion', definicionData.definicion);
+        if(ModoProductivo){
+            const formData = new FormData();
+            formData.append('palabra', definicionData.palabra);
+            formData.append('definicion', definicionData.definicion);
 
-        const respuesta = await axios({
-            url: `${baseURL}/definiciones`,
-            method: 'POST',
-            data: definicionData
-        })
+            const respuesta = await axios({
+                url: `${baseURL}/definiciones`,
+                method: 'POST',
+                data: definicionData
+            })
 
-        return respuesta;
+            return respuesta;
+        }else{
+            definicionesJSON.push(definicionData);
+        }
 
     }catch(e){
         console.log(e)
@@ -72,11 +96,24 @@ export async function cargaDataForm(definicionData){
 export async function actualizaDefinicion(definicionData){
     try{
 
-        const respuesta = await axios({
-            url: `${baseURL}/definiciones/${definicionData.id}`,
-            method: 'PUT',
-            data: definicionData
-        })
+        if(ModoProductivo){
+            const respuesta = await axios({
+                url: `${baseURL}/definiciones/${definicionData.id}`,
+                method: 'PUT',
+                data: definicionData
+            })
+        }else{
+
+            let idAModificar = definicionData.id;
+
+            definicionesJSON.map(function(dato){
+                if(idAModificar== dato.id){
+                    dato.palabra = definicionData.palabra;
+                    dato.definicion = definicionData.definicion;
+                }
+            })
+
+        }
 
     }catch(e){
         console.log("Error al editar: "+e);
@@ -85,11 +122,20 @@ export async function actualizaDefinicion(definicionData){
 
 export async function deleteDefinicion(id){
     try{
-        const respuesta = await axios({
-            url: `${baseURL}/definiciones/${id}`,
-            method: 'DELETE'
-        })
-        return respuesta;
+
+        if(ModoProductivo){
+            const respuesta = await axios({
+                url: `${baseURL}/definiciones/${id}`,
+                method: 'DELETE'
+            })
+            return respuesta;
+
+        }else{
+            let defDelete = definicionesJSON.findIndex((def) => def.id === id);
+            definicionesJSON.splice(defDelete);
+        }
+
+        
     }catch(e){
             console.log(e)
     }
@@ -97,12 +143,27 @@ export async function deleteDefinicion(id){
 
 export async function getDefinicionPorLetra(letra){
     try{
-        const respuesta = await axios({
-            url: `${baseURL}/definiciones/searchLetter/${letra}`,
-            method: 'GET'
-        })
 
-        return respuesta;
+        if(ModoProductivo){
+
+            const respuesta = await axios({
+                url: `${baseURL}/definiciones/searchLetter/${letra}`,
+                method: 'GET'
+            })
+            console.log("res: "+respuesta);
+            return respuesta;
+
+        }else{
+           
+            let defsEncontras = definicionesJSON.filter(x => x.palabra.startsWith(letra));
+
+            let defJSON = JSON.stringify(defsEncontras);
+            let defObj = JSON.parse(defJSON);
+            
+            return defObj;
+        }
+
+        
     }catch(e){
             console.log(e)
     }
